Rename secrets to uppercase
When a secret is added (Settings -> Secrets -> New repository secret) on
GitHub, the secret name is changed to uppercase in the secrets list, even if
it is actually lower/mixed case. I find this confusing.

N.B.: THIS CHANGE REQUIRES REMOVING AND RE-ADDING THE 2 SECRETS USING THE
NEW UPPERCASE NAME!