use a data table for alphafold reference data
because an environment variable does not work on instances
that do not use TPV. this is because  version specific destinations
can not be configured (i.e. version specific environment
variables are impossible)

for admins the change is to add an entry to the data table
for each version

users will see another parameter that they can't change. so
there is no room for user error.